Added new KeyMasterProvider for the API Controller use. Refactored ToggleOn/ToggleOff into the new provider
